Fix model type, Implement loading tracks from project file

+++ library/src/model/frame/frame.rs
@@ -3,8 +3,8 @@ use crate::model::frame::entity::FrameEntity;
 use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct FrameInfo {
-    pub width: u32,
-    pub height: u32,
+    pub width: u64,
+    pub height: u64,
     pub background_color: Color,
     pub color_profile: String,
     pub objects: Vec<FrameEntity>,
